Extract DissertationList as a component to reuse across category and diss pages
